+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 #### Fixed
 
 - MySQLTuner installation
+- `wo stack remove/purge --all`
+- variable substitution in install script
 
 ### v3.9.8 - 2019-08-16